requirements.txt
telethon
hiredis
redis
motor
python-dotenv
pykeyboard
tgcrypto
asyncio
